requirements.txt
alembic==0.8.6
argon2-cffi==16.1.0
cffi==1.7.0
click==6.6
Flask==0.11.1
Flask-Cors==2.1.2
Flask-HTTPAuth==3.1.2
Flask-Migrate==1.8.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.4
MarkupSafe==0.23
marshmallow==2.8.0
psycopg2==2.6.1
pycparser==2.14
python-editor==1.0.1
six==1.10.0
SQLAlchemy==1.0.13
Werkzeug==0.11.10
WTForms==2.1
